FEAT: Adding download permissions to Plato.Articles.Attachments
