+++ src/Web/Modules/Plato.Entities.Attachments/Views/Shared/Components/EntityAttachments/Default.cshtml
@@ -22,7 +22,7 @@
                     {
                         <span class="fiv-sqo @icon align-middle mr-1"></span>
                     }
-                    <a asp-route-area="@attachment.ModuleId" asp-route-controller="Attachment" asp-route-action="Download" asp-route-id="@attachment.AttachmentId" target="_blank" class="align-middle">@attachment.Name</a>
+                    <a asp-route-area="@attachment.ModuleId" asp-route-controller="Home" asp-route-action="Download" asp-route-id="@attachment.AttachmentId" target="_blank" class="align-middle">@attachment.Name</a>
                     <span class="text-muted">@details</span>
                 </div>
             }
