+++ src/Web/Modules/Plato.Entities.Attachments/Views/Shared/Components/PreviewEntityAttachments/Default.cshtml
@@ -11,7 +11,7 @@
                 var tooltip = $"{attachment.Name} - {attachment.ContentLength.ToFriendlyFileSize()}";
                 @if (!string.IsNullOrEmpty(iconCss))
                 {
-                    <a asp-route-area="@attachment.ModuleId" asp-route-controller="Attachment" asp-route-action="Download" asp-route-id="@attachment.Id" data-toggle="tooltip" title="@tooltip" class="d-inline-block mr-1">
+                    <a asp-route-area="@attachment.ModuleId" asp-route-controller="Home" asp-route-action="Download" asp-route-id="@attachment.Id" data-toggle="tooltip" title="@tooltip" target="_blank" class="d-inline-block mr-1">
                         <i class="fiv-sqo @iconCss"></i>
                     </a>
                 }
